Move some creation logic to Vulkan wrappers
